Fix list edge flood vulnerability on negative requested count.
The maximum amount of edges returned from the GraphQL API
is limited, but the <count> is used not only to specify
the number of edges requested, but also to indicate
the direction of loading. Negative <count> numbers were
not tested for the maximum edges limit and so the API
allowed to return any number of edges if negative <count>
was used on a list. This could potentially lead to DOS.

Reported by Petr Bednar.