Add contract license and compiler optimization params.
The contract may be compiled without compiler optimizations, or with
a different number of optimization cycles than is the default value
for Solidity compiler. We need to be able to validate those contracts.
Also, a contract can have an open source license, include
it as an optional input/output param.