Fix missing delegated amount on delegators list os staker.
The total delegated amount must include partial un-delegations
to be consistent with the account tx balances, but SFC does not
provide this amount. We have to load partial withdrawals
from events, pair them with finalized and sum pending withdrawals
manually. This is time and resources consuming operation
so the client side has to keep requested lists count at reasonable
size.