Resolve correct range of proposals before and after contract update.
After the test contract has been deployed and resolved, the ballot
has been updated to include a default option with index 0 (zero)
and name "None". If a ballot does not lead to a clear winner,
this default value will win. But we don't wan it to be displayed
as a valid option voters can opt for.